DEX-671 respond to code review feedback
